server/pyproject.toml
[tool.poetry]
name = "rtdjango"
version = "0.1.0"
description = ""
authors = ["Gabriel Flores <gabriel.flores0503@gmail.com>"]

[tool.black]
line-length = 80

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.2"
djangorestframework = "^3.12.4"
channels = "^3.0.3"
channels-redis = "^3.2.0"
psycopg2-binary = "^2.8.6"
django-model-utils = "^4.1.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pre-commit = "^2.12.1"
flake8 = "^3.9.2"
black = "^21.5b1"
